feat(helper): make is_egyptian_national_id validation and phone validation in other file)
