+++ app.py
@@ -3,7 +3,7 @@ from flask_mysqldb import MySQL
 import json
 import os
 import re
-
+from .helper import is_egyptian_national_id,phone_regex
 app = Flask(__name__)
 
 
@@ -14,18 +14,7 @@ app.config['MYSQL_USER'] =  os.environ.get("MYSQL_USER","myuser")
 app.config['MYSQL_PASSWORD'] = os.environ.get("MYSQL_PASSWORD","mypassword")
 app.config['MYSQL_DB'] = os.environ.get("MYSQL_DB","mydatabase")
 
-phone_regex = r"^(?:\+20|0)?1[0-2]\d{8}$"
 
-def is_egyptian_national_id(national_id):
-    # check if national_id is a string of 14 digits
-    if not isinstance(national_id, str) or not re.match(r'^\d{14}$', national_id):
-        return False
-    
-    # check if the first digit is either 2, 3, 4, or 9
-    if national_id[0] not in ['2', '3', '4', '9']:
-        return False
-
-    return True
 
 
 mysql = MySQL(app)
